refactor(ai-search): refactor search, move chips, add placeholder, ability to stream response'
