+++ src/views/Docs/AlgoSearch.js
@@ -1,16 +1,11 @@
 import {
   Box,
   Button,
-  FormControl,
   InputAdornment,
-  InputLabel,
   Link,
-  MenuItem,
-  Select,
   Stack,
   TextField,
   Typography,
-  useTheme,
 } from '@mui/material';
 import algoliasearch from 'algoliasearch';
 import SearchIcon from '@mui/icons-material/Search';
@@ -31,11 +26,8 @@ import TreeItem, { treeItemClasses } from '@mui/lab/TreeItem';
 import DescriptionIcon from '@mui/icons-material/Description';
 import SubdirectoryArrowRightIcon from '@mui/icons-material/SubdirectoryArrowRight';
 
-import { ArrowBack, ArrowUpward } from '@mui/icons-material';
-import { Prism as SyntaxHighlighter } from 'react-syntax-highlighter';
-import { dracula } from 'react-syntax-highlighter/dist/cjs/styles/prism';
-import { getCookie } from 'cookies-next';
-import { hashMD5 } from 'utils/Md5Hash';
+import { ArrowBack } from '@mui/icons-material';
+import { AiSearch } from 'components/docs/AiSearch';
 
 const SearchBoxMui = ({ currentRefinement, _isSearchStalled, refine }) => (
   <form noValidate action="" role="search">
@@ -63,7 +55,6 @@ const SearchBoxMui = ({ currentRefinement, _isSearchStalled, refine }) => (
 const CustomSearchBox = connectSearchBox(SearchBoxMui);
 
 export const AlgoSearch = () => {
-  const theme = useTheme();
   const [isAiActive, setisAiActive] = React.useState(false);
   const {
     algoliaApiKey: apiKey,
@@ -71,9 +62,6 @@ export const AlgoSearch = () => {
     algoliaIndex: index,
   } = useZestyStore((e) => e);
 
-  const email = getCookie('APP_USER_EMAIL').replace(/%40/g, '@');
-
-  const md5Hash = hashMD5(email.toLowerCase());
   const searchClient = algoliasearch(appId, apiKey);
 
   const indices = [
@@ -90,79 +78,6 @@ export const AlgoSearch = () => {
       name: 'Parsley',
     },
   ];
-
-  const [chatHistory, setChatHistory] = React.useState([
-    {
-      id: 1,
-      type: 'bot',
-      message: 'Hi, How can I help you?',
-    },
-  ]);
-  const [query, setQuery] = React.useState('');
-  const [loading, setLoading] = React.useState(false);
-
-  const askAiHandler = async (e) => {
-    e.preventDefault();
-    setLoading(true);
-    setQuery('');
-    setChatHistory([
-      ...chatHistory,
-      {
-        id: chatHistory.length + 1,
-        type: 'user',
-        message: query.replaceAll('\n', '<br/>'),
-      },
-    ]);
-    try {
-      const resp = await fetch(
-        'https://us-central1-zesty-dev.cloudfunctions.net/aiSearch/query/',
-        {
-          method: 'POST',
-          headers: {
-            'Content-Type': 'application/json',
-          },
-          body: JSON.stringify({ query: `${query} respond in ${language}` }),
-        },
-      );
-
-      const data = await resp.json();
-
-      if (data) {
-        setLoading(false);
-        setChatHistory((prevChatHistory) => [
-          ...prevChatHistory,
-          {
-            id: prevChatHistory.length + 2,
-            type: 'bot',
-            message: data.response.text.replaceAll('\n', '<br/>'),
-            sourceDocuments: data.response.sourceDocuments,
-          },
-        ]);
-      }
-    } catch (error) {
-      console.log(error);
-    }
-  };
-
-  const formatUrlSources = (url) => {
-    const match = url.match(/\/([^/]+)\/?(\?.*)?$/);
-    if (match) {
-      const lastPath = match[1];
-
-      // Replace hyphens with spaces and capitalize each word
-      const formattedText = lastPath
-        .split('-')
-        .map((word) => word.charAt(0).toUpperCase() + word.slice(1))
-        .join(' ');
-
-      return formattedText;
-    }
-  };
-
-  const [language, setLanguage] = React.useState('english');
-  const handleLanguageChange = (e) => {
-    setLanguage(e.target.value);
-  };
 
   return (
     <Stack data-testid="algolia-search-container">
@@ -187,236 +102,7 @@ export const AlgoSearch = () => {
               />
             </Box>
 
-            <Stack sx={{ mt: 4, maxHeight: 500, overflow: 'auto', pb: 2 }}>
-              <Box>
-                {chatHistory.map((item) => {
-                  const message = item.type === 'bot' ? item.message : '';
-                  const lines = message.split('\n');
-                  const indentation = '  ';
-
-                  const indentedResponseText = lines
-                    .map((line) => {
-                      if (line.trim() === '') {
-                        return '';
-                      }
-
-                      return `${indentation}${line}`;
-                    })
-                    .join('\n');
-
-                  return (
-                    <>
-                      {item.type === 'bot' ? (
-                        <>
-                          <Box
-                            key={item.id}
-                            sx={{
-                              display: 'flex',
-                              gap: 1.5,
-                              mt: 2,
-                            }}
-                          >
-                            <Box>
-                              <Box
-                                component={'img'}
-                                src="https://kfg6bckb.media.zestyio.com/content.one-logo.png"
-                                sx={{
-                                  width: 40,
-                                  height: 35,
-                                  borderRadius: 1,
-                                  display: 'flex',
-                                }}
-                              />
-                            </Box>
-
-                            <Box sx={{ pr: 2 }}>
-                              <MuiMarkdown
-                                options={{
-                                  overrides: {
-                                    code: {
-                                      component: SyntaxHighlighter,
-                                      props: {
-                                        style: dracula,
-                                        language: 'html',
-                                        wrapLongLines: true,
-                                      },
-                                    },
-                                  },
-                                }}
-                              >
-                                {message}
-                              </MuiMarkdown>
-                            </Box>
-                          </Box>
-
-                          <Box
-                            sx={{
-                              ml: 4,
-                              mt: 1,
-                            }}
-                          >
-                            {item.sourceDocuments && (
-                              <>
-                                <Typography
-                                  sx={{
-                                    px: 1,
-                                    py: 0.5,
-                                    borderRadius: 100,
-                                    color: theme.palette.zesty.zestyZambezi,
-                                  }}
-                                  variant="caption"
-                                >
-                                  Sources:
-                                </Typography>
-                                <Box
-                                  sx={{
-                                    mt: 1,
-                                    display: 'flex',
-                                    gap: 1,
-                                    flexWrap: 'wrap',
-                                  }}
-                                >
-                                  {item?.sourceDocuments?.map((item) => {
-                                    return (
-                                      <Box>
-                                        <Box
-                                          component="a"
-                                          href={item.metadata.source}
-                                          target="_blank"
-                                          sx={{
-                                            background: `${theme.palette.zesty.zestyOrange}`,
-                                            borderRadius: 100,
-                                            px: 1,
-                                            display: 'flex',
-                                            alignItems: 'center',
-                                            textDecoration: 'none',
-                                          }}
-                                        >
-                                          <Typography
-                                            sx={{
-                                              color: '#fff',
-                                            }}
-                                            variant="caption"
-                                          >
-                                            {formatUrlSources(
-                                              item.metadata.source,
-                                            )}
-                                          </Typography>
-                                        </Box>
-                                      </Box>
-                                    );
-                                  })}
-                                </Box>
-                              </>
-                            )}
-                          </Box>
-                        </>
-                      ) : (
-                        <>
-                          <Box
-                            key={item.id}
-                            sx={{
-                              display: 'flex',
-                              justifyContent: 'flex-end',
-                              mt: 2,
-                            }}
-                          >
-                            <Box
-                              sx={{
-                                display: 'flex',
-                                gap: 2,
-                                pl: 5,
-                              }}
-                            >
-                              <Box sx={{ mt: 0.5 }}>
-                                <MuiMarkdown options={{ wrapper: 'p' }}>
-                                  {item.message.replaceAll(/\n/g, '<br/>')}
-                                </MuiMarkdown>
-                              </Box>
-
-                              <Box
-                                component={'img'}
-                                src={`https://www.gravatar.com/avatar/${md5Hash}?s=200&d=mm`}
-                                sx={{
-                                  width: 40,
-                                  height: 40,
-                                  borderRadius: 1,
-                                  display: 'block',
-                                }}
-                              />
-                            </Box>
-                          </Box>
-                        </>
-                      )}
-                    </>
-                  );
-                })}
-              </Box>
-            </Stack>
-
-            <Box
-              sx={{
-                display: 'flex',
-                justifyContent: 'center',
-                alignItems: 'center',
-                mt: 4,
-              }}
-            >
-              {loading && (
-                <Typography
-                  variant="caption"
-                  sx={{ color: theme.palette.zesty.zestyZambezi }}
-                >
-                  Searching documentation. This may take a second!
-                </Typography>
-              )}
-            </Box>
-
-            <Box
-              sx={{ display: 'flex', gap: 1, mt: 1 }}
-              component="form"
-              onSubmit={askAiHandler}
-            >
-              <TextField
-                value={query}
-                onChange={(e) => setQuery(e.target.value)}
-                sx={{ flex: 1 }}
-              ></TextField>
-              <Button
-                type="submit"
-                startIcon={<ArrowUpward />}
-                variant="contained"
-                color="secondary"
-              >
-                Ask
-              </Button>
-
-              <Box sx={{ width: 80 }}>
-                <FormControl fullWidth>
-                  <InputLabel>Lang</InputLabel>
-                  <Select
-                    value={language}
-                    label="Select Language"
-                    onChange={handleLanguageChange}
-                  >
-                    <MenuItem value={'english'}>EN</MenuItem>
-                    <MenuItem value={'spanish'}>ES</MenuItem>
-                    <MenuItem value={'tagalog'}>TL</MenuItem>
-                    <MenuItem value={'hindi'}>HI</MenuItem>
-                    <MenuItem value={'italian'}>IT</MenuItem>
-                    <MenuItem value={'french'}>FR</MenuItem>
-                    <MenuItem value={'german'}>DE</MenuItem>
-                  </Select>
-                </FormControl>
-              </Box>
-            </Box>
-            <Typography
-              sx={{ textAlign: 'center', mt: 1, color: '#ccc' }}
-              variant="caption"
-            >
-              Content.one AI Chatbot powered by ChatGPT may produce inaccurate
-              Parsley code.
-            </Typography>
+            <AiSearch />
           </>
         ) : (
           <>
@@ -440,7 +126,7 @@ export const AlgoSearch = () => {
             <Stack sx={{ mt: 2 }} maxHeight={'50vh'} overflow={'auto'}>
               {indices.map((item) => {
                 return (
-                  <Index indexName={item.title}>
+                  <Index key={item.title} indexName={item.title}>
                     <Box sx={{ px: 2, py: 2 }}>
                       <Typography
                         data-testid={`${item.title}-index`}
